Changed datatype to BigInteger (using int internally where the concern is only an individual combinational value; which makes a limit of 11! )

+++ Datafield/Datafield.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using combgen.Datatype;
 
 namespace combgen.Datafield;
@@ -10,7 +11,7 @@ public abstract class Datafield
         Default,
         Verbose
     };
-    public abstract DataType Read(int combVal, int? aIndex, int? bIndex);
+    public abstract DataType Read(BigInteger combVal, int? aIndex, int? bIndex);
     public abstract int Count();
-    public abstract string GetTable(int baseIndex, TableVerbosity tv = TableVerbosity.Default, string title = "Value");
+    public abstract string GetTable(BigInteger baseIndex, TableVerbosity tv = TableVerbosity.Default, string title = "Value");
 }
